Update taskpane.js
Created line for adding to localStorage in the addToLib function.

Created an allStorage function that pulls all keys from local storage.

Started but did not completed the showLibraryRemote function which may replace the showInLib function or may go inside said function.

Created a removeFromLocal function which takes a key and wipes that from localStorage and then added a function call for that in the showLibrary function.